fix(yalb-984): move getGlobalThemes to new file, import that into components
