+++ components/01-atoms/controls/control.stories.js
@@ -1,4 +1,5 @@
 import tokens from '@yalesites-org/tokens/build/json/tokens.json';
+import getGlobalThemes from '../../00-tokens/colors/color-global-themes';
 import ctaTwig from './cta/yds-cta.twig';
 import linkTwig from './text-link/yds-text-link.twig';
 
@@ -9,16 +10,7 @@ import themeExamplesTwig from './cta/_yds-cta-examples.twig';
 const siteGlobalThemes = { themes: tokens['global-themes'] };
 const componentThemes = { themes: tokens['component-themes'] };
 const componentThemeOptions = Object.keys(tokens['component-themes']);
-
-const globalThemeLabels = {
-  themes: {
-    'One - Old Blues': 'one',
-    'Two - New Haven Green': 'two',
-    'Three - Shoreline Summer': 'three',
-    'Four - Elm City Nights': 'four',
-    'Five - Quiet Corner': 'five',
-  },
-};
+const siteGlobalThemeOptions = getGlobalThemes(tokens['global-themes']);
 
 /**
  * Storybook Definition.
@@ -322,7 +314,7 @@ export const CtaExamples = ({ globalTheme, componentTheme }) =>
 CtaExamples.argTypes = {
   globalTheme: {
     name: 'Global Theme (lever)',
-    options: globalThemeLabels.themes,
+    options: siteGlobalThemeOptions,
     type: 'select',
     defaultValue: 'one',
   },
