+++ components/03-organisms/site-footer/site-footer.stories.js
@@ -1,4 +1,5 @@
 import tokens from '@yalesites-org/tokens/build/json/tokens.json';
+import getGlobalThemes from '../../00-tokens/colors/color-global-themes';
 
 import siteFooterTwig from './yds-site-footer.twig';
 import siteFooterExamples from './_site-footer--examples.twig';
@@ -9,16 +10,7 @@ const siteFooterThemes = { themes: tokens['site-footer-themes'] };
 const siteGlobalThemes = { themes: tokens['global-themes'] };
 const borderThicknessOptions = Object.keys(tokens.border.thickness);
 const siteFooterThemeOptions = Object.keys(tokens['site-footer-themes']);
-
-const globalThemeLabels = {
-  themes: {
-    'One - Old Blues': 'one',
-    'Two - New Haven Green': 'two',
-    'Three - Shoreline Summer': 'three',
-    'Four - Elm City Nights': 'four',
-    'Five - Quiet Corner': 'five',
-  },
-};
+const siteGlobalThemeOptions = getGlobalThemes(tokens['global-themes']);
 
 /**
  * Storybook Definition.
@@ -64,7 +56,7 @@ export const FooterExamples = ({ borderThickness, globalTheme }) =>
 FooterExamples.argTypes = {
   globalTheme: {
     name: 'Global Theme (lever)',
-    options: globalThemeLabels.themes,
+    options: siteGlobalThemeOptions,
     type: 'select',
     defaultValue: 'one',
   },
