+++ components/_settings/config.js
@@ -37,15 +37,4 @@ const setAttributes = (dataAttributes) => {
   }
 };
 
-const getGlobalThemes = (globalThemeTokens) => {
-  const globalThemeOptions = {};
-  const tempArr = Object.keys(globalThemeTokens);
-
-  tempArr.forEach((element) => {
-    globalThemeOptions[globalThemeTokens[element].label] = element;
-  });
-
-  return globalThemeOptions;
-};
-
-export { setAttributes, getGlobalThemes };
+export default setAttributes;
