+++ components/_settings/config.stories.js
@@ -1,5 +1,6 @@
 import tokens from '@yalesites-org/tokens/build/json/tokens.json';
-import { setAttributes, getGlobalThemes } from './config';
+import setAttributes from './config';
+import getGlobalThemes from '../00-tokens/colors/color-global-themes';
 
 // Twig files.
 import configTwig from './config.twig';
@@ -15,6 +16,7 @@ const widths = Object.keys(tokens.layout.width);
 const borderThicknessOptions = Object.keys(tokens.border.thickness);
 const siteHeaderThemeOptions = Object.keys(tokens['site-header-themes']);
 const siteFooterThemeOptions = Object.keys(tokens['site-footer-themes']);
+const siteGlobalThemeOptions = getGlobalThemes(tokens['global-themes']);
 
 export default {
   title: 'Config',
@@ -24,7 +26,7 @@ export default {
   argTypes: {
     globalTheme: {
       name: 'Global Theme (lever)',
-      options: getGlobalThemes(tokens['global-themes']),
+      options: siteGlobalThemeOptions,
       type: 'select',
       defaultValue: localStorage.getItem('yds-cl-twig-global-theme'),
     },
